+++ src/main/kotlin/app/revanced/patches/nova/prime/annotations/UnlockPrimeCompatibility.kt
@@ -1,13 +1,13 @@
-package app.revanced.patches.reddit.ad.general.annotations
+package app.revanced.patches.nova.prime.annotations
 
 import app.revanced.patcher.annotation.Compatibility
 import app.revanced.patcher.annotation.Package
 
 @Compatibility(
-    [Package(
-        "com.reddit.frontpage", arrayOf("2022.43.0")
-    )]
+    [
+        Package("com.teslacoilsw.launcher")
+    ]
 )
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
-internal annotation class GeneralAdsCompatibility
+internal annotation class UnlockPrimeCompatibility